Permission to modify and delete, Display member's nickname in articles.

+++ src/main/java/com/NonContact/controller/UsrArticleController.java
@@ -3,6 +3,8 @@ package com.NonContact.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.NonContact.dto.Article;
 import com.NonContact.dto.ResultData;
 import com.NonContact.service.ArticleService;
+import com.NonContact.util.Util;
 
 @Controller
 public class UsrArticleController {
@@ -20,15 +23,23 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/detail")
 	@ResponseBody
-	public Article showDetail(int id) {
-		Article article = articleService.getArticle(id);
+	public ResultData showDetail(Integer id) {
+		
+		if(id == null) {
+			return new ResultData("F-2", "id를 입력해주세요.");
+		}
+		Article article = articleService.getForPrintArticle(id);
+		
+		if(article == null) {
+			return new ResultData("F-1", "해당 게시물이 존재하지 않습니다.");
+		}
 
-		return article;
+		return new ResultData("S-1", "성공", "article", article);
 	}
 
 	@RequestMapping("/usr/article/list")
 	@ResponseBody
-	public List<Article> showList(String searchKeywordType, String searchKeyword) {
+	public ResultData showList(String searchKeywordType, String searchKeyword) {
 		if (searchKeywordType != null) {
 			searchKeywordType = searchKeywordType.trim();
 		}
@@ -45,25 +56,40 @@ public class UsrArticleController {
 			searchKeywordType = null;
 		}
 
-		return articleService.getArticles(searchKeywordType, searchKeyword);
+		List<Article> articles = articleService.getForPrintArticles(searchKeywordType, searchKeyword);
+
+		return new ResultData("S-1", "성공", "articles", articles);
 	}
 
 	// 게시물 추가
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData doAdd(@RequestParam Map<String, Object> param) {
+	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginedMemberId == 0) {
+			return new ResultData("F-5", "로그인 후 이용가능합니다.");
+		}
+
 		if (param.get("title") == null && param.get("body") == null) {
 			return new ResultData("F-2", "title 또는 body가 입력되지 않았습니다.");
 		}
-		ResultData rsData = articleService.addArticle(param);
 
-		return rsData;
+		param.put("memberId", loginedMemberId);
+
+		return articleService.addArticle(param);
 	}
 
 	// 게시물 삭제
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginedMemberId == 0) {
+			return new ResultData("F-5", "로그인 후 이용가능합니다.");
+		}
 
 		if (id == null) {
 			return new ResultData("F-2", "id를 입력해주세요.");
@@ -73,13 +99,26 @@ public class UsrArticleController {
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물이 존재하지 않습니다.");
 		}
+
+		ResultData AuthChkRd = articleService.getAuthChkRd(article, loginedMemberId);
+
+		if (AuthChkRd.isFail()) {
+			return AuthChkRd;
+		}
+
 		return articleService.deleteArticle(id);
 	}
 
 	// 게시물 수정
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+
+		if (loginedMemberId == 0) {
+			return new ResultData("F-5", "로그인 후 이용가능합니다.");
+		}
 
 		if (id == null) {
 			return new ResultData("F-2", "id를 입력해주세요.");
@@ -92,6 +131,13 @@ public class UsrArticleController {
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물이 존재하지 않습니다.");
 		}
+
+		ResultData AuthChkRd = articleService.getAuthChkRd(article, loginedMemberId);
+
+		if (AuthChkRd.isFail()) {
+			return AuthChkRd;
+		}
+
 		return articleService.modifyArticle(id, body, title);
 	}
